+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Procedure when bumping the version number:
 - Authentication dialog for password-protected rooms
 - Error popups in rooms when something goes wrong
 - `--ephemeral` flag that prevents cove from storing data permanently
+- Key binding to download more logs
 
 ### Changed
 - Reduced amount of unnecessary redraws